+++ src/Line.h
@@ -22,10 +22,9 @@ class Line : public GeneralGeometryElement {
   int start_point;  ///< Index of the start point
   int end_point;    ///< Index of the end point
   std::array<std::array<double, DIM>, LINE_DIM>
-      corners;                  ///< Array of line corners
-  std::array<double, DIM> out;  ///< Output point of the line
-  std::array<double, DIM - LINE_DIM>
-      const_i;  ///< Constant indices for the line
+      corners;                              ///< Array of line corners
+  std::array<double, DIM> out;              ///< Output point of the line
+  std::array<int, DIM - LINE_DIM> const_i;  ///< Constant indices for the line
 
  public:
   /**
@@ -52,7 +51,7 @@ class Line : public GeneralGeometryElement {
    */
   void init_line(std::array<std::array<double, DIM>, LINE_DIM> new_corners,
                  std::array<double, DIM> new_out,
-                 std::array<double, DIM - LINE_DIM> new_const_i);
+                 std::array<int, DIM - LINE_DIM> new_const_i);
 
   /**
    * @brief Flips the start and end points of the line.
